app/requirements.txt
asyncio-nats-client
sqlalchemy
sqlalchemy-utils
psycopg2
apispec
marshmallow
marshmallow-sqlalchemy
flask
watchdog
